Revert "Entidad movimiento"
