+++ routes/pokemones/pokemon.py
@@ -1,14 +1,6 @@
 from fastapi import HTTPException, APIRouter
-from db import (
-    lista_pokemones,
-    fortalezas_tipos,
-    debilidades_tipos,
-    datos_pokemon,
-    datos_movimientos_pokemon,
-    datos_movimientos,
-    datos_tipos_pokemon,
-)
-from models import Pokemon, PokemonCreate, Evolucion, Movimientomoves
+from db import lista_pokemones, fortalezas_tipos, debilidades_tipos, Pokemon, PokemonCreate, datos_pokemon, datos_movimientos_pokemon, datos_movimientos, datos_tipos_pokemon, Evolucion, Movimientomoves
+
 
 router = APIRouter()
 
@@ -89,7 +81,7 @@ def create_pokemon(pokemon: PokemonCreate):
         estadisticas=pokemon.estadisticas,
         habilidades=pokemon.habilidades,
         generaciones=pokemon.generaciones,
-        evoluciones_inmediatas=pokemon.evoluciones_inmediatas,
+        evoluciones_inmediatas=pokemon.evoluciones_inmediatas
     )
     lista_pokemones.append(nuevo_pokemon)
     return nuevo_pokemon
@@ -120,14 +112,7 @@ def obtener_movimientos_pokemon(pokemon_id: int):
             movimientos_vistos.add(id_movimiento)
             if id_movimiento in datos_movimientos.movimientos:
                 nombre_movimiento = datos_movimientos.movimientos[id_movimiento].nombre
-                lista_movimientos.append(
-                    Movimientomoves(
-                        id=id_movimiento,
-                        nombre=nombre_movimiento,
-                        nivel=nivel_movimiento,
-                        es_evolucionado=False,
-                    )
-                )
+                lista_movimientos.append(Movimientomoves(id=id_movimiento, nombre=nombre_movimiento, nivel=nivel_movimiento, es_evolucionado=False))
 
     tipos = datos_tipos_pokemon.get(pokemon_id, [])
