+++ routes/teams/teams.py
@@ -1,21 +1,14 @@
-from models import *
-from fastapi import APIRouter, HTTPException, status
-from sqlmodel import select
-
+from fastapi import HTTPException, APIRouter
+from typing import List
 from db import (
-    SessionDep,
     lista_naturalezas,
     lista_pokemones,
     lista_movimientos,
     lista_habilidades,
-    generaciones_pokemon,
-    pokemon_tipos,
-)
-from fastapi import HTTPException, APIRouter
-from typing import List
-from models import (
     TeamCreate,
     TeamDataCreate,
+    generaciones_pokemon,
+    pokemon_tipos,
     Pokemon,
 )
